+++ modules/utils/biorxiv_api.py
@@ -1,0 +1,13 @@
+import time
+from typing import List
+
+from modules.utils.query import get_value
+
+
+def process_data(json_info: dict, section: str, keys: List[str], cursor: int, disable: bool = True) -> List:
+    journal_list = [[entry + cursor] + [get_value(journal, key) for key in keys] for entry, journal in
+                    enumerate(json_info[section])]
+    if disable is False:
+        time.sleep(60)
+
+    return journal_list
